+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 *.vsix
 scripts/
 **/.DS_Store
+coverage/**
