+++ jest.config.js
@@ -2,4 +2,5 @@
 module.exports = {
     // Other Jest configuration options...
     testPathIgnorePatterns: ['/bin/'],
+    collectCoverage: true,
 };
